requirements.txt
opencv-python
paddlepaddle
paddleocr
ollama
TTS
ffmpeg-python
PyYAML
pydantic
tqdm
PyMuPDF
requests
typing_extensions
